[B] Combined vs Combining conditionally render
Previously the condition for rendeirng Combined vs Combining page was
checking to see if the page had been visited before.  This was always
triggering true because the visitedPages global state value was
evaluated immediately before/as the page loaded.  Now it checks to see
if the user has visited any page > the page in question.

After much deliberating: used this as an opprotunity to refactor Global
state stuff into its own class, and instantiate that class, and call
some of its class functions in index.jsx before rendering App component
so state is already available.